+++ example/ios/Runner/AppDelegate.swift
@@ -1,5 +1,6 @@
-import UIKit
 import Flutter
+import UIKit
+import home_widget
 import workmanager
 
 @UIApplicationMain
@@ -9,13 +10,19 @@ import workmanager
     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
   ) -> Bool {
     GeneratedPluginRegistrant.register(with: self)
-    
-    UIApplication.shared.setMinimumBackgroundFetchInterval(TimeInterval(60*15))
-    
+
+    UIApplication.shared.setMinimumBackgroundFetchInterval(TimeInterval(60 * 15))
+
     WorkmanagerPlugin.setPluginRegistrantCallback { registry in
-        GeneratedPluginRegistrant.register(with: registry)
+      GeneratedPluginRegistrant.register(with: registry)
     }
-    
+
+    if #available(iOS 17, *) {
+      HomeWidgetBackgroundWorker.setPluginRegistrantCallback { registry in
+        GeneratedPluginRegistrant.register(with: registry)
+      }
+    }
+
     return super.application(application, didFinishLaunchingWithOptions: launchOptions)
   }
 }