+++ example/ios/Runner/BackgroundIntent.swift
@@ -1,0 +1,40 @@
+//
+//  BackgroundIntent.swift
+//  Runner
+//
+//  Created by Anton Borries on 26.08.23.
+//
+
+import AppIntents
+import Foundation
+import home_widget
+
+@available(iOS 17, *)
+public struct BackgroundIntent: AppIntent {
+  static public var title: LocalizedStringResource = "HomeWidget Background Intent"
+
+  @Parameter(title: "Widget URI")
+  var url: URL?
+
+  @Parameter(title: "AppGroup")
+  var appGroup: String?
+
+  public init() {}
+
+  public init(url: URL?, appGroup: String?) {
+    self.url = url
+    self.appGroup = appGroup
+  }
+
+  public func perform() async throws -> some IntentResult {
+    await HomeWidgetBackgroundWorker.run(url: url, appGroup: appGroup!)
+
+    return .result()
+  }
+}
+
+/// This is required if you want to have the widget be interactive even when the app is fully suspended.
+/// Note that this will launch your App so on the Flutter side you should check for the current Lifecycle State before doing heavy tasks
+@available(iOS 17, *)
+@available(iOSApplicationExtension, unavailable)
+extension BackgroundIntent: ForegroundContinuableIntent {}