+++ lib/home_widget.dart
@@ -97,9 +97,20 @@ class HomeWidget {
   }
 
   /// Register a callback that gets called when clicked on a specific View in a HomeWidget
-  /// supported only on Android
+  /// This enables having Interactive Widgets that can call Dart Code
   /// More Info on setting this up in the README
-  static Future<bool?> registerBackgroundCallback(Function(Uri?) callback) {
+  @Deprecated('Use `registerInteractivityCallback` instead')
+  static Future<bool?> registerBackgroundCallback(
+    FutureOr<void> Function(Uri?) callback,
+  ) =>
+      registerInteractivityCallback(callback);
+
+  /// Register a callback that gets called when clicked on a specific View in a HomeWidget
+  /// This enables having Interactive Widgets that can call Dart Code
+  /// More Info on setting this up in the README
+  static Future<bool?> registerInteractivityCallback(
+    FutureOr<void> Function(Uri?) callback,
+  ) {
     final args = <dynamic>[
       ui.PluginUtilities.getCallbackHandle(callbackDispatcher)?.toRawHandle(),
       ui.PluginUtilities.getCallbackHandle(callback)?.toRawHandle(),